+++ Carbon.IIS/Functions/Set-CIisCollectionItem.ps1
@@ -5,20 +5,23 @@ function Set-CIisCollectionItem
     Adds a new item to an IIS configuration collection.
 
     .DESCRIPTION
-    The `Set-CIisCollectionItem` function adds a new item to an IIS configuration collection. Pass the collection's IIS
-    configuration section path to the `SectionPath` parameter and the value to add to the collection to the `Value`
-    parameter. The function adds that value to the collection if it doesn't already exist in the collection or if its
-    value has changed.
+    The `Set-CIisCollectionItem` function adds a new item or updates the configuration of an existing IIS configuration
+    collection item. Pipe the item value to the function, or pass it to the `InputObject` parameter. If the collection
+    items you're configuring have only one attribute/value, pass just the value. Otherwise, pass a hashtable of
+    attribute names/values.
 
-    To add extra attributes to the collection item, pass the *extra* attributes as a hashtable to the `Attribute`
-    parameter. If the item already exists in the collection, any attributes on the item that aren't in the `Attribute`
-    hasthable are *removed* from the collection item.
+    To configure a collection that is part of a global configuration section, pass the configuration section's path to
+    the `SectionPath` parameter. If the configuration section itself isn't a collection, pass the name of the collection
+    to the `Name` parameter. To configure a configuration section for a specific site, directory, application, or
+    virtual directory, pass its location path to the `LocationPath` parameter. To configure a specific
+    `[Microsoft.Web.Administration.ConfigurationElement]` item (i.e. a site, application pool, etc.), pass that object
+    to the `ConfigurationElement` parameter. If the configuration element itself isn't a collection, pass the name of
+    the object' collection property to the `Name` parameter.
 
-    If the configuration section given by `SectionPath` isn't a collection, pass the name of the collection to the
-    `CollectionName` parameter.
-
-    If adding an item to the collection for a website, application, virtual directory, or directory, pass that path to
-    that location to the `LocationPath` parameter.
+    When making changes directly to ConfigurationElement objects, test that those changes are saved correctly to the IIS
+    application host configuration. Some configuration has to be saved at the same time as its parent configuration
+    elements are created (i.e. sites, application pools, etc.). Use the `Suspend-CIisAutoCommit` and
+    `Resume-CIisAutoCommit` functions to ensure configuration gets committed simultaneously.
 
     .EXAMPLE
     Add-CIisCollectionItem -SectionPath 'system.webServer/httpProtocol/customHeaders' -Value 'X-Api-Header'
@@ -56,139 +59,169 @@ function Set-CIisCollectionItem
     #>
     [CmdletBinding(DefaultParameterSetName='Global')]
     param(
+        [Parameter(Mandatory, ParameterSetName='Direct')]
+        [ConfigurationElement] $ConfigurationElement,
+
         # The site to add the item to
         [Parameter(Mandatory, ParameterSetName='Location')]
         [String] $LocationPath,
 
         # The path for the configuration section to edit
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, ParameterSetName='Global')]
+        [Parameter(Mandatory, ParameterSetName='Location')]
         [String] $SectionPath,
 
         # The value for the IIS collection's identifying key.
-        [Parameter(Mandatory)]
-        [Object] $Value,
+        [Parameter(Mandatory, ValueFromPipeline)]
+        [Object] $InputObject,
 
         # The name of the IIS collection to modify. If not provided, will use the SectionPath as the collection.
-        [String] $CollectionName,
-
-        # Additional attributes to be added to a new IIS item.
-        [AllowNull()]
-        [hashtable] $Attribute
+        [String] $CollectionName
     )
-    Set-StrictMode -Version 'Latest'
-    Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    $save = $false
-
-    $msgs = [System.Collections.ArrayList] @()
-
-    $collectionArgs = @{}
-
-    $msgPrefix = 'IIS section '
-    $outputSection = "$($LocationPath):$($SectionPath)"
-
-    if ($LocationPath)
+    begin
     {
-        $msgPrefix = $msgPrefix + "$($LocationPath)/"
-        $collectionArgs['LocationPath'] = $LocationPath
-    }
+        Set-StrictMode -Version 'Latest'
+        Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    $msgPrefix = $msgPrefix + "$($SectionPath)"
+        $save = $false
 
-    if ($CollectionName)
-    {
-        $msgPrefix = $msgPrefix + "/$($CollectionName)"
-        $outputSection + "/$($CollectionName)"
-        $collectionArgs['Name'] = $CollectionName
-    }
+        $msgs = [System.Collections.ArrayList] @()
 
-    [void]$msgs.Add($msgPrefix)
+        $collectionArgs = @{}
 
-    $collection = Get-CIisCollection @collectionArgs -SectionPath $SectionPath
+        $msgPrefix = 'IIS configuration collection '
 
-    if (-not $collection)
-    {
-        return
-    }
-
-    $keyAttrName = Get-CIisCollectionKeyName -Collection $collection
-
-    if (-not $keyAttrName)
-    {
-        $msg = "Unable to find key for $($outputSection)"
-        Write-Error -Message $msg
-        return
-    }
-
-    $add = $collection | Where-Object { $_.GetAttributeValue($keyAttrName) -eq $Value }
-
-    if (-not $add)
-    {
-        $add = $collection.CreateElement('add')
-        $add.SetAttributeValue($keyAttrName, $Value)
-        [void]$collection.Add($add)
-        [void]$msgs.Add("  + ${keyAttrName} = ${Value}")
-        $save = $true
-    }
-    else
-    {
-        [void]$msgs.Add("    ${keyattrname} = ${value}")
-    }
-
-    if (-not $attribute)
-    {
-        $attribute = @{}
-    }
-
-    foreach ($attrname in $attribute.keys)
-    {
-        if ($attrname -eq $keyattrname)
+        $elementPath = ''
+        if ($ConfigurationElement)
         {
-            $msg = "Pass walue with key ""$($attrname)"" as the Value parameter. Attribute is only for extra attributes"
-            Write-Warning -Message $msg
-            continue
+            $msgPrefix = "${msgPrefix}$($ConfigurationElement.ElementTagName)"
+            $collectionArgs['ConfigurationElement'] = $ConfigurationElement
+            $elementPath = $ConfigurationElement.ElementTagName
+            if (Get-Member -Name 'SectionPath' -InputObject $ConfigurationElement)
+            {
+                $elementPath = $ConfigurationElement.SectionPath
+            }
+        }
+        else
+        {
+            $displayPath = Get-CIisDisplayPath -SectionPath $SectionPath `
+                                               -LocationPath $locationPath `
+                                               -SubSectionPath $CollectionName
+            $msgPrefix = "${msgPrefix}${displayPath}"
+            $elementPath = $SectionPath
+            $collectionArgs['SectionPath'] = $SectionPath
+            if ($LocationPath)
+            {
+                $collectionArgs['LocationPath'] = $LocationPath
+            }
         }
 
-        $expectedValue = $attribute[$attrname]
-        $actualValue = $add.getattributevalue($attrname)
-
-        if ($null -eq $actualValue)
+        if ($CollectionName)
         {
-            [void]$msgs.Add("      + ${attrname} = ${expectedValue}")
-            $add.setattributevalue($attrname, $expectedValue)
+            $elementPath = "${elementPath}/$($CollectionName)"
+            $collectionArgs['Name'] = $CollectionName
+        }
+
+        [void]$msgs.Add($msgPrefix)
+
+        $collection = Get-CIisCollection @collectionArgs
+
+        if (-not $collection)
+        {
+            return
+        }
+
+        $keyAttrName = Get-CIisCollectionKeyName -Collection $collection
+
+        if (-not $keyAttrName)
+        {
+            $msg = "Unable to find key for ${elementPath}."
+            Write-Error -Message $msg
+            return
+        }
+
+    }
+
+    process
+    {
+        if (-not $collection)
+        {
+            return
+        }
+
+        if ($InputObject -isnot [IDictionary])
+        {
+            $InputObject = @{ $keyAttrName = $InputObject }
+        }
+
+        if (-not $InputObject.Contains($keyAttrName))
+        {
+            $msg = "Failed to add item to collection ""$($collection.Path)"" because the attributes of the item are " +
+                   "missing a value for the key attribute ""${keyAttrName}""."
+            Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+            return
+        }
+
+        $keyValue = $InputObject[$keyAttrName]
+
+        $item = $collection | Where-Object { $_.GetAttributeValue($keyAttrName) -eq $keyValue }
+
+        if (-not $item)
+        {
+            $item = $collection.CreateElement('add')
+            foreach ($attrName in $InputObject.Keys)
+            {
+                $item.SetAttributeValue($attrName, $InputObject[$attrName])
+            }
+            [void]$collection.Add($item)
+            [void]$msgs.Add("  + ${keyAttrName} = ${keyValue}")
             $save = $true
-            continue
-        }
-        elseif ($expectedValue -ne $actualValue)
-        {
-            [void]$msgs.Add("      ${attrname}  $($actualValue) -> ${expectedValue}")
-            $add.setattributevalue($attrname, $expectedValue)
-            $save = $true
-        }
-    }
-
-    foreach ($attr in $add.Attributes)
-    {
-        if ($attr.Name -eq $keyAttrName)
-        {
-            continue
         }
 
-        if (-not $Attribute.containskey($attr.Name))
+        foreach ($attrName in $InputObject.Keys)
         {
+            $expectedValue = $InputObject[$attrName]
+            $actualValue = $item.GetAttributeValue($attrName)
+
+            if ($null -eq $actualValue)
+            {
+                [void]$msgs.Add("      + ${attrName} = ${expectedValue}")
+                $item.SetAttributeValue($attrName, $expectedValue)
+                $save = $true
+                continue
+            }
+            elseif ($expectedValue -ne $actualValue)
+            {
+                [void]$msgs.Add("      ${attrName}  $($actualValue) -> ${expectedValue}")
+                $item.SetAttributeValue($attrName, $expectedValue)
+                $save = $true
+            }
+        }
+
+        foreach ($attr in $item.Attributes)
+        {
+            if ($InputObject.Contains($attr.Name))
+            {
+                continue
+            }
+
             [void]$msgs.Add("    - $($attr.Name)")
-            [void]$attr.delete()
+            [void]$attr.Delete()
             $save = $true
         }
     }
 
-    if ($save)
+    end
     {
-        foreach ($msg in $msgs)
+        if ($save)
         {
-            Write-Information $msg
-        }
+            foreach ($msg in $msgs)
+            {
+                Write-Information $msg
+            }
 
-        Save-CIisConfiguration
+            Save-CIisConfiguration
+        }
     }
 }