+++ Carbon.IIS/Functions/Stop-CIisWebsite.ps1
@@ -82,6 +82,10 @@ function Stop-CIisWebsite
                 }
                 catch
                 {
+                    if ($script:skipCommit)
+                    {
+                        return
+                    }
                     $lastError = $_
                     Start-Sleep -Milliseconds 100
                     $website = Get-CIisWebsite -Name $website.Name